fix: wire proxy rule set export/import to the live page; remove dead components
The export/import feature (#332) was added to ProxyRuleSetsTab.tsx, which is
dead code — it's only re-exported by an orphan barrel and rendered nowhere.
The Proxy Rules tab at /repo/:owner/:repo/proxy-rules is served by
ProxyRuleSetsPage.tsx, so the feature never appeared in the UI.

- Port the Export menu item, Import button, and export/import handlers
  (including secret-value stripping on export) to ProxyRuleSetsPage.tsx.
- Remove the dead cluster: ProxyRuleSetsTab.tsx, RuleSetEditorDialog.tsx,
  ProxyRulesEditor.tsx, and the orphan components/proxy-rules/index.ts barrel.
  (ProxyRuleForm.tsx is kept — still used by ExpandedProxyRuleForm/RuleEditorPage.)

Backend import endpoint from #332 is unchanged and already live.
