--- apps/frontend/src/components/proxy-rules/ProxyRuleSetsTab.tsx
@@ -1,523 +1,0 @@
-import { useRef, useState } from 'react';
-import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Skeleton } from '@/components/ui/skeleton';
-import { Badge } from '@/components/ui/badge';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
-import {
-  Plus,
-  MoreHorizontal,
-  Edit2,
-  Trash2,
-  Settings,
-  Layers,
-  CopyPlus,
-  Download,
-  Upload,
-} from 'lucide-react';
-import {
-  useGetProjectRuleSetsQuery,
-  useDeleteRuleSetMutation,
-  useCopyRuleSetMutation,
-  useLazyGetRuleSetQuery,
-  useImportRuleSetMutation,
-  type ProxyRuleSetExport,
-  type ExportedProxyRule,
-  type ProxyRuleSetWithRules,
-  type HeaderConfig,
-} from '@/services/proxyRulesApi';
-import { useGetProjectQuery } from '@/services/projectsApi';
-import { useProjectRole } from '@/hooks/useProjectRole';
-import { useToast } from '@/hooks/use-toast';
-import { CreateRuleSetDialog } from './CreateRuleSetDialog';
-import { RuleSetEditorDialog } from './RuleSetEditorDialog';
-
-interface ProxyRuleSetsTabProps {
-  owner: string;
-  repo: string;
-}
-
-/**
- * Header `add` values can hold secrets (API keys, bearer tokens). We deliberately
- * do NOT export their values — exports keep the header names but blank the values,
- * so the user re-enters secrets after import. This is a known export/import gap
- * until inline secrets move to project-level `secrets.<name>` references
- * (see stories/tasks/migrate-inline-pipeline-secrets-to-project-secrets.md).
- */
-function sanitizeHeaderConfigForExport(headerConfig: HeaderConfig | null): HeaderConfig | null {
-  if (!headerConfig?.add) return headerConfig;
-  const blankedAdd: Record<string, string> = {};
-  for (const key of Object.keys(headerConfig.add)) {
-    blankedAdd[key] = '';
-  }
-  return { ...headerConfig, add: blankedAdd };
-}
-
-/**
- * ProxyRuleSetsTab - Displays and manages proxy rule sets for a repository
- * Used in RepositoryOverviewPage on the Proxy Rules tab
- */
-export function ProxyRuleSetsTab({ owner, repo }: ProxyRuleSetsTabProps) {
-  const { toast } = useToast();
-  const { canEdit } = useProjectRole(owner, repo);
-  const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
-  const [editingRuleSet, setEditingRuleSet] = useState<{
-    id: string;
-    name: string;
-  } | null>(null);
-  const [deletingRuleSet, setDeletingRuleSet] = useState<{
-    id: string;
-    name: string;
-  } | null>(null);
-
-  // Fetch project to get projectId
-  const { data: project, isLoading: isLoadingProject } = useGetProjectQuery({ owner, name: repo });
-
-  // Fetch rule sets for this project
-  const {
-    data: ruleSetsData,
-    isLoading: isLoadingRuleSets,
-    error: ruleSetsError,
-  } = useGetProjectRuleSetsQuery(project?.id || '', {
-    skip: !project?.id,
-  });
-
-  // Delete mutation
-  const [deleteRuleSet, { isLoading: isDeleting }] = useDeleteRuleSetMutation();
-
-  // Copy mutation
-  const [copyRuleSet, { isLoading: isCopying }] = useCopyRuleSetMutation();
-
-  // Export / import
-  const [fetchRuleSet, { isFetching: isExporting }] = useLazyGetRuleSetQuery();
-  const [importRuleSet, { isLoading: isImporting }] = useImportRuleSetMutation();
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const handleExport = async (ruleSetId: string) => {
-    try {
-      const ruleSet = await fetchRuleSet(ruleSetId).unwrap();
-      const exportData: ProxyRuleSetExport = {
-        version: 1,
-        exportedAt: new Date().toISOString(),
-        kind: 'bffless-proxy-rule-set',
-        ruleSet: {
-          name: ruleSet.name,
-          description: ruleSet.description,
-          environment: ruleSet.environment,
-        },
-        // Strip server-managed fields; keep only portable rule config
-        rules: ruleSet.rules.map(
-          (rule): ExportedProxyRule => ({
-            pathPattern: rule.pathPattern,
-            method: rule.method,
-            targetUrl: rule.targetUrl,
-            stripPrefix: rule.stripPrefix,
-            order: rule.order,
-            timeout: rule.timeout,
-            preserveHost: rule.preserveHost,
-            forwardCookies: rule.forwardCookies,
-            headerConfig: sanitizeHeaderConfigForExport(rule.headerConfig),
-            authTransform: rule.authTransform,
-            internalRewrite: rule.internalRewrite,
-            proxyType: rule.proxyType,
-            emailHandlerConfig: rule.emailHandlerConfig,
-            pipelineConfig: rule.pipelineConfig,
-            isEnabled: rule.isEnabled,
-            debugEnabled: rule.debugEnabled,
-            description: rule.description,
-          }),
-        ),
-      };
-
-      const blob = new Blob([JSON.stringify(exportData, null, 2)], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-      const link = document.createElement('a');
-      const safeName = ruleSet.name.replace(/[^a-zA-Z0-9-_]+/g, '-').replace(/^-+|-+$/g, '');
-      link.href = url;
-      link.download = `${safeName || 'proxy-rule-set'}.proxy-rules.json`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      URL.revokeObjectURL(url);
-
-      const hadSecrets = ruleSet.rules.some(
-        (rule) => rule.headerConfig?.add && Object.keys(rule.headerConfig.add).length > 0,
-      );
-      toast({
-        title: 'Rule set exported',
-        description:
-          `Exported "${ruleSet.name}" with ${ruleSet.rules.length} rule${ruleSet.rules.length !== 1 ? 's' : ''}.` +
-          (hadSecrets ? ' Added-header secret values were not included.' : ''),
-      });
-    } catch {
-      toast({
-        title: 'Export failed',
-        description: 'Could not export the rule set. Please try again.',
-        variant: 'destructive',
-      });
-    }
-  };
-
-  const handleImportClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleImportFile = async (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    // Reset the input so selecting the same file again re-triggers onChange
-    event.target.value = '';
-    if (!file || !project?.id) return;
-
-    let parsed: ProxyRuleSetExport;
-    try {
-      parsed = JSON.parse(await file.text());
-    } catch {
-      toast({
-        title: 'Invalid file',
-        description: 'The selected file is not valid JSON.',
-        variant: 'destructive',
-      });
-      return;
-    }
-
-    if (!parsed?.ruleSet?.name || !Array.isArray(parsed.rules)) {
-      toast({
-        title: 'Invalid file',
-        description: 'This file is not a valid proxy rule set export.',
-        variant: 'destructive',
-      });
-      return;
-    }
-
-    try {
-      const imported = (await importRuleSet({
-        projectId: project.id,
-        data: parsed,
-      }).unwrap()) as ProxyRuleSetWithRules;
-      toast({
-        title: 'Rule set imported',
-        description: `Created "${imported.name}" with ${imported.rules.length} rule${imported.rules.length !== 1 ? 's' : ''}.`,
-      });
-    } catch (err: unknown) {
-      const errorMessage =
-        (err as { data?: { message?: string } })?.data?.message || 'Failed to import rule set';
-      toast({
-        title: 'Import failed',
-        description: errorMessage,
-        variant: 'destructive',
-      });
-    }
-  };
-
-  const handleCopy = async (ruleSetId: string) => {
-    try {
-      const copiedRuleSet = await copyRuleSet(ruleSetId).unwrap();
-      toast({
-        title: 'Rule set copied',
-        description: `Created "${copiedRuleSet.name}" with ${copiedRuleSet.rules.length} rule${copiedRuleSet.rules.length !== 1 ? 's' : ''}.`,
-      });
-    } catch (err: unknown) {
-      const errorMessage =
-        (err as { data?: { message?: string } })?.data?.message || 'Failed to copy rule set';
-      toast({
-        title: 'Error',
-        description: errorMessage,
-        variant: 'destructive',
-      });
-    }
-  };
-
-  const handleDelete = async () => {
-    if (!deletingRuleSet) return;
-    try {
-      await deleteRuleSet(deletingRuleSet.id).unwrap();
-      toast({
-        title: 'Rule set deleted',
-        description: `Rule set "${deletingRuleSet.name}" has been deleted.`,
-      });
-      setDeletingRuleSet(null);
-    } catch (err: unknown) {
-      const errorMessage =
-        (err as { data?: { message?: string } })?.data?.message || 'Failed to delete rule set';
-      toast({
-        title: 'Error',
-        description: errorMessage,
-        variant: 'destructive',
-      });
-    }
-  };
-
-  const isLoading = isLoadingProject || isLoadingRuleSets;
-
-  // Shared header action buttons (Import + Create), reused across states
-  const headerActions = canEdit ? (
-    <div className="flex items-center gap-2">
-      <Button
-        variant="outline"
-        size="sm"
-        className="gap-2"
-        onClick={handleImportClick}
-        disabled={isImporting || !project?.id}
-      >
-        <Upload className="h-4 w-4" />
-        {isImporting ? 'Importing...' : 'Import'}
-      </Button>
-      <Button size="sm" className="gap-2" onClick={() => setIsCreateDialogOpen(true)}>
-        <Plus className="h-4 w-4" />
-        Create Rule Set
-      </Button>
-    </div>
-  ) : null;
-
-  // Hidden file input used by the Import button
-  const importFileInput = (
-    <input
-      ref={fileInputRef}
-      type="file"
-      accept="application/json,.json"
-      className="hidden"
-      onChange={handleImportFile}
-    />
-  );
-
-  // Loading state
-  if (isLoading) {
-    return (
-      <Card>
-        <CardHeader>
-          <CardTitle>Proxy Rule Sets</CardTitle>
-          <CardDescription>Manage reusable proxy rule configurations</CardDescription>
-        </CardHeader>
-        <CardContent>
-          <div className="space-y-4">
-            <div className="flex justify-end">
-              <Skeleton className="h-10 w-40" />
-            </div>
-            {[...Array(3)].map((_, i) => (
-              <Skeleton key={i} className="h-24 w-full" />
-            ))}
-          </div>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  // Error state
-  if (ruleSetsError) {
-    return (
-      <Card>
-        <CardHeader>
-          <CardTitle>Proxy Rule Sets</CardTitle>
-        </CardHeader>
-        <CardContent className="p-8 text-center">
-          <p className="text-destructive">Failed to load proxy rule sets</p>
-          <Button
-            variant="outline"
-            size="sm"
-            className="mt-4"
-            onClick={() => window.location.reload()}
-          >
-            Retry
-          </Button>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  const ruleSets = ruleSetsData?.ruleSets || [];
-
-  // Empty state
-  if (ruleSets.length === 0) {
-    return (
-      <>
-        {importFileInput}
-        <Card>
-          <CardHeader className="flex flex-row items-center justify-between">
-            <div>
-              <CardTitle>Proxy Rule Sets</CardTitle>
-              <CardDescription>Manage reusable proxy rule configurations</CardDescription>
-            </div>
-            {canEdit && (
-              <Button size="sm" className="gap-2" onClick={() => setIsCreateDialogOpen(true)}>
-                <Plus className="h-4 w-4" />
-                Create Rule Set
-              </Button>
-            )}
-          </CardHeader>
-          <CardContent className="p-8 text-center">
-            <Layers className="h-12 w-12 mx-auto text-muted-foreground mb-4" />
-            <p className="text-muted-foreground font-medium">No proxy rule sets found</p>
-            <p className="text-sm text-muted-foreground mt-2">
-              {canEdit
-                ? 'Create a rule set to define proxy rules that can be assigned to aliases'
-                : 'No proxy rule sets have been created for this repository yet'}
-            </p>
-          </CardContent>
-        </Card>
-
-        {canEdit && (
-          <CreateRuleSetDialog
-            projectId={project?.id || ''}
-            open={isCreateDialogOpen}
-            onOpenChange={setIsCreateDialogOpen}
-          />
-        )}
-      </>
-    );
-  }
-
-  return (
-    <>
-      {importFileInput}
-      <Card>
-        <CardHeader className="flex flex-row items-center justify-between">
-          <div>
-            <CardTitle>Proxy Rule Sets</CardTitle>
-            <CardDescription>Manage reusable proxy rule configurations</CardDescription>
-          </div>
-          {headerActions}
-        </CardHeader>
-        <CardContent>
-          {/* Rule Sets List */}
-          <div className="space-y-3">
-            {ruleSets.map((ruleSet) => (
-              <div
-                key={ruleSet.id}
-                className="flex items-center justify-between p-4 rounded-lg border hover:bg-accent transition-colors"
-              >
-                <div className="flex items-center gap-4">
-                  <Settings className="h-5 w-5 text-muted-foreground" />
-                  <div>
-                    <div className="flex items-center gap-2">
-                      <span className="font-medium">{ruleSet.name}</span>
-                      {ruleSet.environment && (
-                        <Badge variant="outline" className="text-xs">
-                          {ruleSet.environment}
-                        </Badge>
-                      )}
-                      {project?.defaultProxyRuleSetId === ruleSet.id && (
-                        <Badge variant="secondary" className="text-xs">
-                          Default
-                        </Badge>
-                      )}
-                    </div>
-                    {ruleSet.description && (
-                      <p className="text-sm text-muted-foreground mt-1">{ruleSet.description}</p>
-                    )}
-                  </div>
-                </div>
-
-                {canEdit && (
-                  <div className="flex items-center gap-2">
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      className="gap-1"
-                      onClick={() => setEditingRuleSet({ id: ruleSet.id, name: ruleSet.name })}
-                    >
-                      <Edit2 className="h-3 w-3" />
-                      Edit Rules
-                    </Button>
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" size="icon">
-                          <MoreHorizontal className="h-4 w-4" />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <DropdownMenuContent align="end">
-                        <DropdownMenuItem
-                          onClick={() => handleCopy(ruleSet.id)}
-                          disabled={isCopying}
-                        >
-                          <CopyPlus className="h-4 w-4 mr-2" />
-                          {isCopying ? 'Duplicating...' : 'Duplicate'}
-                        </DropdownMenuItem>
-                        <DropdownMenuItem
-                          onClick={() => handleExport(ruleSet.id)}
-                          disabled={isExporting}
-                        >
-                          <Download className="h-4 w-4 mr-2" />
-                          {isExporting ? 'Exporting...' : 'Export'}
-                        </DropdownMenuItem>
-                        <DropdownMenuItem
-                          onClick={() => setDeletingRuleSet({ id: ruleSet.id, name: ruleSet.name })}
-                          className="text-destructive"
-                        >
-                          <Trash2 className="h-4 w-4 mr-2" />
-                          Delete
-                        </DropdownMenuItem>
-                      </DropdownMenuContent>
-                    </DropdownMenu>
-                  </div>
-                )}
-              </div>
-            ))}
-          </div>
-        </CardContent>
-      </Card>
-
-      {/* Create Dialog */}
-      {canEdit && (
-        <CreateRuleSetDialog
-          projectId={project?.id || ''}
-          open={isCreateDialogOpen}
-          onOpenChange={setIsCreateDialogOpen}
-        />
-      )}
-
-      {/* Edit Rules Dialog */}
-      {canEdit && editingRuleSet && (
-        <RuleSetEditorDialog
-          ruleSetId={editingRuleSet.id}
-          ruleSetName={editingRuleSet.name}
-          open={!!editingRuleSet}
-          onOpenChange={(open) => !open && setEditingRuleSet(null)}
-        />
-      )}
-
-      {/* Delete Confirmation Dialog */}
-      {canEdit && (
-        <AlertDialog
-          open={!!deletingRuleSet}
-          onOpenChange={(open) => !open && setDeletingRuleSet(null)}
-        >
-          <AlertDialogContent>
-            <AlertDialogHeader>
-              <AlertDialogTitle>Delete Rule Set</AlertDialogTitle>
-              <AlertDialogDescription>
-                Are you sure you want to delete the rule set "{deletingRuleSet?.name}"? This will
-                remove all rules in this set and unassign it from any aliases using it. This action
-                cannot be undone.
-              </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-              <AlertDialogCancel disabled={isDeleting}>Cancel</AlertDialogCancel>
-              <AlertDialogAction
-                onClick={handleDelete}
-                className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-                disabled={isDeleting}
-              >
-                {isDeleting ? 'Deleting...' : 'Delete'}
-              </AlertDialogAction>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialog>
-      )}
-    </>
-  );
-}
