--- apps/frontend/src/components/proxy-rules/index.ts
@@ -1,8 +1,0 @@
-export { ProxyRulesEditor } from './ProxyRulesEditor';
-export { ProxyRuleForm } from './ProxyRuleForm';
-export { ProxyRuleSetsTab } from './ProxyRuleSetsTab';
-export { CreateRuleSetDialog } from './CreateRuleSetDialog';
-export { RuleSetEditorDialog } from './RuleSetEditorDialog';
-export { RuleSetCard } from './RuleSetCard';
-export { RulesList } from './RulesList';
-export { ExpandedProxyRuleForm } from './ExpandedProxyRuleForm';
