+++ apps/frontend/src/pages/ProxyRuleSetsPage.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
@@ -19,14 +19,40 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-import { Plus, Layers, MoreHorizontal, Trash2, CopyPlus } from 'lucide-react';
-import { useGetProjectRuleSetsQuery, useDeleteRuleSetMutation, useCopyRuleSetMutation } from '@/services/proxyRulesApi';
+import { Plus, Layers, MoreHorizontal, Trash2, CopyPlus, Download, Upload } from 'lucide-react';
+import {
+  useGetProjectRuleSetsQuery,
+  useDeleteRuleSetMutation,
+  useCopyRuleSetMutation,
+  useLazyGetRuleSetQuery,
+  useImportRuleSetMutation,
+  type ProxyRuleSetExport,
+  type ExportedProxyRule,
+  type ProxyRuleSetWithRules,
+  type HeaderConfig,
+} from '@/services/proxyRulesApi';
 import { useGetProjectQuery } from '@/services/projectsApi';
 import { useProjectRole } from '@/hooks/useProjectRole';
 import { useToast } from '@/hooks/use-toast';
 import { CreateRuleSetDialog } from '@/components/proxy-rules/CreateRuleSetDialog';
 import { RuleSetCard } from '@/components/proxy-rules/RuleSetCard';
 import { routes } from '@/utils/routes';
+
+/**
+ * Header `add` values can hold secrets (API keys, bearer tokens). We deliberately
+ * do NOT export their values — exports keep the header names but blank the values,
+ * so the user re-enters secrets after import. This is a known export/import gap
+ * until inline secrets move to project-level `secrets.<name>` references
+ * (see stories/tasks/migrate-inline-pipeline-secrets-to-project-secrets.md).
+ */
+function sanitizeHeaderConfigForExport(headerConfig: HeaderConfig | null): HeaderConfig | null {
+  if (!headerConfig?.add) return headerConfig;
+  const blankedAdd: Record<string, string> = {};
+  for (const key of Object.keys(headerConfig.add)) {
+    blankedAdd[key] = '';
+  }
+  return { ...headerConfig, add: blankedAdd };
+}
 
 /**
  * ProxyRuleSetsPage - Content for the Proxy Rules tab showing all rule sets.
@@ -63,6 +89,127 @@ export function ProxyRuleSetsPage() {
 
   // Copy mutation
   const [copyRuleSet, { isLoading: isCopying }] = useCopyRuleSetMutation();
+
+  // Export / import
+  const [fetchRuleSet, { isFetching: isExporting }] = useLazyGetRuleSetQuery();
+  const [importRuleSet, { isLoading: isImporting }] = useImportRuleSetMutation();
+  const fileInputRef = useRef<HTMLInputElement>(null);
+
+  const handleExport = async (ruleSetId: string) => {
+    try {
+      const ruleSet = await fetchRuleSet(ruleSetId).unwrap();
+      const exportData: ProxyRuleSetExport = {
+        version: 1,
+        exportedAt: new Date().toISOString(),
+        kind: 'bffless-proxy-rule-set',
+        ruleSet: {
+          name: ruleSet.name,
+          description: ruleSet.description,
+          environment: ruleSet.environment,
+        },
+        // Strip server-managed fields; keep only portable rule config
+        rules: ruleSet.rules.map(
+          (rule): ExportedProxyRule => ({
+            pathPattern: rule.pathPattern,
+            method: rule.method,
+            targetUrl: rule.targetUrl,
+            stripPrefix: rule.stripPrefix,
+            order: rule.order,
+            timeout: rule.timeout,
+            preserveHost: rule.preserveHost,
+            forwardCookies: rule.forwardCookies,
+            headerConfig: sanitizeHeaderConfigForExport(rule.headerConfig),
+            authTransform: rule.authTransform,
+            internalRewrite: rule.internalRewrite,
+            proxyType: rule.proxyType,
+            emailHandlerConfig: rule.emailHandlerConfig,
+            pipelineConfig: rule.pipelineConfig,
+            isEnabled: rule.isEnabled,
+            debugEnabled: rule.debugEnabled,
+            description: rule.description,
+          }),
+        ),
+      };
+
+      const blob = new Blob([JSON.stringify(exportData, null, 2)], { type: 'application/json' });
+      const url = URL.createObjectURL(blob);
+      const link = document.createElement('a');
+      const safeName = ruleSet.name.replace(/[^a-zA-Z0-9-_]+/g, '-').replace(/^-+|-+$/g, '');
+      link.href = url;
+      link.download = `${safeName || 'proxy-rule-set'}.proxy-rules.json`;
+      document.body.appendChild(link);
+      link.click();
+      document.body.removeChild(link);
+      URL.revokeObjectURL(url);
+
+      const hadSecrets = ruleSet.rules.some(
+        (rule) => rule.headerConfig?.add && Object.keys(rule.headerConfig.add).length > 0,
+      );
+      toast({
+        title: 'Rule set exported',
+        description:
+          `Exported "${ruleSet.name}" with ${ruleSet.rules.length} rule${ruleSet.rules.length !== 1 ? 's' : ''}.` +
+          (hadSecrets ? ' Added-header secret values were not included.' : ''),
+      });
+    } catch {
+      toast({
+        title: 'Export failed',
+        description: 'Could not export the rule set. Please try again.',
+        variant: 'destructive',
+      });
+    }
+  };
+
+  const handleImportClick = () => {
+    fileInputRef.current?.click();
+  };
+
+  const handleImportFile = async (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files?.[0];
+    // Reset the input so selecting the same file again re-triggers onChange
+    event.target.value = '';
+    if (!file || !project?.id) return;
+
+    let parsed: ProxyRuleSetExport;
+    try {
+      parsed = JSON.parse(await file.text());
+    } catch {
+      toast({
+        title: 'Invalid file',
+        description: 'The selected file is not valid JSON.',
+        variant: 'destructive',
+      });
+      return;
+    }
+
+    if (!parsed?.ruleSet?.name || !Array.isArray(parsed.rules)) {
+      toast({
+        title: 'Invalid file',
+        description: 'This file is not a valid proxy rule set export.',
+        variant: 'destructive',
+      });
+      return;
+    }
+
+    try {
+      const imported = (await importRuleSet({
+        projectId: project.id,
+        data: parsed,
+      }).unwrap()) as ProxyRuleSetWithRules;
+      toast({
+        title: 'Rule set imported',
+        description: `Created "${imported.name}" with ${imported.rules.length} rule${imported.rules.length !== 1 ? 's' : ''}.`,
+      });
+    } catch (err: unknown) {
+      const errorMessage =
+        (err as { data?: { message?: string } })?.data?.message || 'Failed to import rule set';
+      toast({
+        title: 'Import failed',
+        description: errorMessage,
+        variant: 'destructive',
+      });
+    }
+  };
 
   const handleCopy = async (ruleSetId: string) => {
     try {
@@ -151,6 +298,13 @@ export function ProxyRuleSetsPage() {
 
   return (
     <>
+      <input
+        ref={fileInputRef}
+        type="file"
+        accept="application/json,.json"
+        className="hidden"
+        onChange={handleImportFile}
+      />
       <Card>
         <CardHeader className="flex flex-row items-center justify-between">
           <div>
@@ -158,10 +312,22 @@ export function ProxyRuleSetsPage() {
             <CardDescription>Manage reusable proxy rule configurations</CardDescription>
           </div>
           {canEdit && (
-            <Button size="sm" className="gap-2" onClick={() => setIsCreateDialogOpen(true)}>
-              <Plus className="h-4 w-4" />
-              Create Rule Set
-            </Button>
+            <div className="flex items-center gap-2">
+              <Button
+                variant="outline"
+                size="sm"
+                className="gap-2"
+                onClick={handleImportClick}
+                disabled={isImporting || !project?.id}
+              >
+                <Upload className="h-4 w-4" />
+                {isImporting ? 'Importing...' : 'Import'}
+              </Button>
+              <Button size="sm" className="gap-2" onClick={() => setIsCreateDialogOpen(true)}>
+                <Plus className="h-4 w-4" />
+                Create Rule Set
+              </Button>
+            </div>
           )}
         </CardHeader>
         <CardContent>
@@ -209,6 +375,16 @@ export function ProxyRuleSetsPage() {
                           >
                             <CopyPlus className="h-4 w-4 mr-2" />
                             {isCopying ? 'Duplicating...' : 'Duplicate'}
+                          </DropdownMenuItem>
+                          <DropdownMenuItem
+                            onClick={(e) => {
+                              e.stopPropagation();
+                              handleExport(ruleSet.id);
+                            }}
+                            disabled={isExporting}
+                          >
+                            <Download className="h-4 w-4 mr-2" />
+                            {isExporting ? 'Exporting...' : 'Export'}
                           </DropdownMenuItem>
                           <DropdownMenuItem
                             onClick={(e) => {
